[#420] Implement custom procApiRequest for forward/backward compatibility
This commit introduces an implementation of procApiRequest which allows
the HTTP API to handle forward and backward compatibility issues
stemming from differences in packing instructions.

The motivating use-case is iRODS 5's introduction of access time
tracking. This enhancement to the iRODS server caused the structure of
DataObjInfo to grow a new member. The new member caused the KeyValPair
member of ModDataObjInp to shift in a backward incompatible way.