[#460] Return 400 on failure to open replica for reading
Returning a HTTP status code of 500 can lead users to believe the HTTP
server or iRODS server experienced an internal error. Users interacting
with the HTTP API are more likely to pass incorrect arguments to the
"read" operation, hence this commit.

Until the dstream library is able to return error codes, the best we can
do is to have the user verify the arguments passed are correct.